requirements.txt
websocket-client>=1.6.0
requests>=2.28.0
bcrypt>=4.0.0
